AlgoOrdersHistoryButton component
