+++ src/components/AlgoOrdersHistoryButton/index.js
@@ -1,0 +1,1 @@
+export { default } from './AlgoOrdersHistoryButton'
