+++ src/components/AlgoOrdersHistoryButton/style.scss
@@ -1,0 +1,3 @@
+.hfui-history-button__icon {
+  padding-top: 6px;
+}
