+++ src/components/AlgoOrdersTable/style.scss
@@ -20,7 +20,3 @@
     }
   }
 }
-
-.history-icon {
-  padding-top: 6px;
-}
