+++ src/components/TradingStatePanel/TradingStatePanel.container.js
@@ -9,11 +9,9 @@ import {
 } from '../../redux/selectors/ws'
 import { getMarkets } from '../../redux/selectors/meta'
 import UIActions from '../../redux/actions/ui'
-import AOActions from '../../redux/actions/ao'
 import { getComponentState, getCurrentMode } from '../../redux/selectors/ui'
 
 import TradingStatePanel from './TradingStatePanel'
-import { getShowAOsHistory } from '../../redux/selectors/ao'
 
 const mapStateToProps = (state = {}, { layoutID, layoutI: id } = {}) => ({
   authToken: getAuthToken(state),
@@ -24,7 +22,6 @@ const mapStateToProps = (state = {}, { layoutID, layoutI: id } = {}) => ({
   markets: getMarkets(state),
   savedState: getComponentState(state, layoutID, 'trading_state', id),
   getCurrencySymbol: reduxSelectors.getCurrencySymbolMemo(state),
-  isAOsHistoryActive: getShowAOsHistory(state),
 })
 
 const mapDispatchToProps = (dispatch) => ({
@@ -37,7 +34,6 @@ const mapDispatchToProps = (dispatch) => ({
       }),
     )
   },
-  toggleShowAOsHistory: (currentState) => dispatch(AOActions.setShowAOsHistory(!currentState)),
 })
 
 export default connect(mapStateToProps, mapDispatchToProps)(TradingStatePanel)
